Splits Profile and Games pages
